+++ documents/docs/zh-CN/changelog.md
@@ -31,6 +31,8 @@
 
 [优化] 内容超出显示 tips 支持宽度设置-适应列宽 [内容超出设置](#/zh-CN/component/basic/overflow)
 
+[优化] 右键菜单 支持自适应方向显示 需要更新 `soulTable.js soulTable.css`
+
 ### **1.3.27** <small>`2019-09-25`</small>
 
 [文档] 修复后台分页的示例报错问题